Half points (0.5) allowed, using double instead of int (except entities and DB)

+++ src/main/java/exambyte/application/dto/ReviewDTO.java
@@ -7,4 +7,4 @@ public record ReviewDTO (
         UUID antwortId,
         UUID korrektorId,
         String bewertung,
-        int punkte) {}
+        double punkte) {}
